add `ui/slider` tests
